adds hiera.yaml tiers
With this addition, a customer can use either pp_apptier trusted data OR
an external fact (evpn_vxlan_role) when configuring a switch to assign as
spine or leaf.

This doesn't use the pp_role trusted data, because that role may be site
specific, using pp_apptier is the most logical since either this node is
a spine or a leaf.

If the user doesn't want to use trusted data (regen certs, etc), they can
also add evpn_vxlan_role = spine OR leaf to /etc/puppetlabs/fact/facts.d